// contracts/typechain-types/DropCashEscrow.ts
export interface DropCashEscrowInterface extends Interface {
  getFunction(
    nameOrSignature: "cancel" | "claim" | "create" | "escrows" | "getEscrow"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "Cancelled" | "Claimed" | "Created"
  ): EventFragment;

  encodeFunctionData(functionFragment: "cancel", values: [BytesLike]): string;
  encodeFunctionData(
    functionFragment: "claim",
    values: [BytesLike, string, AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "create",
    values: [BytesLike, BytesLike, BigNumberish]
  ): string;
  encodeFunctionData(functionFragment: "escrows", values: [BytesLike]): string;
  encodeFunctionData(
    functionFragment: "getEscrow",
    values: [BytesLike]
  ): string;

  decodeFunctionResult(functionFragment: "cancel", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "claim", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "create", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "escrows", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getEscrow", data: BytesLike): Result;
}

export namespace CancelledEvent {
  export type InputTuple = [
    linkId: BytesLike,
    sender: AddressLike,
    amount: BigNumberish
  ];
  export type OutputTuple = [linkId: string, sender: string, amount: bigint];
  export interface OutputObject {
    linkId: string;
    sender: string;
    amount: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace ClaimedEvent {
  export type InputTuple = [
    linkId: BytesLike,
    recipient: AddressLike,
    amount: BigNumberish
  ];
  export type OutputTuple = [linkId: string, recipient: string, amount: bigint];
  export interface OutputObject {
    linkId: string;
    recipient: string;
    amount: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace CreatedEvent {
  export type InputTuple = [
    linkId: BytesLike,
    sender: AddressLike,
    amount: BigNumberish,
    expiresAt: BigNumberish
  ];
  export type OutputTuple = [
    linkId: string,
    sender: string,
    amount: bigint,
    expiresAt: bigint
  ];
  export interface OutputObject {
    linkId: string;
    sender: string;
    amount: bigint;
    expiresAt: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface DropCashEscrow extends BaseContract {
  connect(runner?: ContractRunner | null): DropCashEscrow;
  waitForDeployment(): Promise<this>;

  interface: DropCashEscrowInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  cancel: TypedContractMethod<[linkId: BytesLike], [void], "nonpayable">;

  claim: TypedContractMethod<
    [linkId: BytesLike, code: string, recipient: AddressLike],
    [void],
    "nonpayable"
  >;

  create: TypedContractMethod<
    [linkId: BytesLike, codeHash: BytesLike, expiresAt: BigNumberish],
    [void],
    "payable"
  >;

  escrows: TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, bigint, string, boolean, boolean, bigint] & {
        sender: string;
        amount: bigint;
        codeHash: string;
        claimed: boolean;
        cancelled: boolean;
        expiresAt: bigint;
      }
    ],
    "view"
  >;

  getEscrow: TypedContractMethod<
    [linkId: BytesLike],
    [
      [string, bigint, boolean, boolean, bigint] & {
        sender: string;
        amount: bigint;
        claimed: boolean;
        cancelled: boolean;
        expiresAt: bigint;
      }
    ],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "cancel"
  ): TypedContractMethod<[linkId: BytesLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "claim"
  ): TypedContractMethod<
    [linkId: BytesLike, code: string, recipient: AddressLike],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "create"
  ): TypedContractMethod<
    [linkId: BytesLike, codeHash: BytesLike, expiresAt: BigNumberish],
    [void],
    "payable"
  >;
  getFunction(
    nameOrSignature: "escrows"
  ): TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, bigint, string, boolean, boolean, bigint] & {
        sender: string;
        amount: bigint;
        codeHash: string;
        claimed: boolean;
        cancelled: boolean;
        expiresAt: bigint;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "getEscrow"
  ): TypedContractMethod<
    [linkId: BytesLike],
    [
      [string, bigint, boolean, boolean, bigint] & {
        sender: string;
        amount: bigint;
        claimed: boolean;
        cancelled: boolean;
        expiresAt: bigint;
      }
    ],
    "view"
  >;

  getEvent(
    key: "Cancelled"
  ): TypedContractEvent<
    CancelledEvent.InputTuple,
    CancelledEvent.OutputTuple,
    CancelledEvent.OutputObject
  >;
  getEvent(
    key: "Claimed"
  ): TypedContractEvent<
    ClaimedEvent.InputTuple,
    ClaimedEvent.OutputTuple,
    ClaimedEvent.OutputObject
  >;
  getEvent(
    key: "Created"
  ): TypedContractEvent<
    CreatedEvent.InputTuple,
    CreatedEvent.OutputTuple,
    CreatedEvent.OutputObject
  >;

  filters: {
    "Cancelled(bytes32,address,uint256)": TypedContractEvent<
      CancelledEvent.InputTuple,
      CancelledEvent.OutputTuple,
      CancelledEvent.OutputObject
    >;
    Cancelled: TypedContractEvent<
      CancelledEvent.InputTuple,
      CancelledEvent.OutputTuple,
      CancelledEvent.OutputObject
    >;

    "Claimed(bytes32,address,uint256)": TypedContractEvent<
      ClaimedEvent.InputTuple,
      ClaimedEvent.OutputTuple,
      ClaimedEvent.OutputObject
    >;
    Claimed: TypedContractEvent<
      ClaimedEvent.InputTuple,
      ClaimedEvent.OutputTuple,
      ClaimedEvent.OutputObject
    >;

    "Created(bytes32,address,uint256,uint256)": TypedContractEvent<
      CreatedEvent.InputTuple,
      CreatedEvent.OutputTuple,
      CreatedEvent.OutputObject
    >;
    Created: TypedContractEvent<
      CreatedEvent.InputTuple,
      CreatedEvent.OutputTuple,
      CreatedEvent.OutputObject
    >;
  };
}